feat: Extracts helper fns
- Adds validation check to ensure a `false` boolean value does not get added to the class list.
